Data.Path.Lens

Traversals

class ToTraversal (l :: * -> * -> *) where

Methods

toTraversal :: l a b -> Traversal' a b

pathed :: (IsPath p, ToTraversal l) => p l a b -> Traversal' a b

Helpers

newtype T a b

Constructors

T

Fields

unT :: Traversal' a b

Ixed

newtype I (p :: (* -> * -> *) -> * -> * -> *) (l :: * -> * -> *) (a :: *) (b :: *)

I p l a b wraps an a such that a path p l a b can be used as an index to traverse a b

ix :: p l a b -> Traversal' (I p l a b) b
ix = pathed

Constructors

I a